Refactor auth handling and share SSH_MSG_USERAUTH_BANNER handling.
