+++ src/Tmds.Ssh/Constants.cs
@@ -14,4 +14,5 @@ static class Constants
     public const int MaxNameLength = 128;        // Arbitrary limit, may be increased.
     public const int MaxECPointLength = 256 + 1; // Arbitrary limit, may be increased.
     public const int MaxKeyLength = 1024;        // Arbitrary limit, may be increased.
+    public const int MaxBannerPackets = 1024;    // Abitrary limit
 }
